requirements.txt
djangorestframework
django
django-unfold
django-modeltranslation
drf-spectacular
